lib/qemu.py: Parse --pexpect-timeout as a string
In pexpect library, a timeout of 0 can be used to specify no timeout
mode where we want to wait indefinitely for commands.  However, passing
the timeout as 0 currently is broken because integer 0 is interpreted as
a NULL value and hence a defualt timeout of 60s is used. Change the
--pexpect-timeout to be interpreted as a string so that we can pass 0
as well.

Signed-off-by: Ojaswin Mujoo <ojaswin@linux.ibm.com>