lib/qemu.py: Add support to pass shared output dir
This output dir is shared between the guest running the test and the
host. The test runner can use this shared dir to store logs of the test
back to the host.

Further, add support for collecting logs to tests/avocado/test.py
runner. This can be referred to for reference usage.

Signed-off-by: Ojaswin Mujoo <ojaswin@linux.ibm.com>