build/scripts: Add CUSTOM_CONFIG option to pass custom kernel config
Currently, the custom .configs are supported via DEFCONFIG=".config*"
(check build/scripts/container-build.sh for details), however this
support is broken as there is no mechanism to pass the custom config
from the host to the docker container. Further, passing it in DEFCONFIG
is confusing.

Enable support for a CUSTOM_CONFIG options which takes absolute path
of the config on host and exposes it as /config.txt in the container.
The container-build.sh script running inside the container can then
copy it from there.

Signed-off-by: Ojaswin Mujoo <ojaswin@linux.ibm.com>